refactor: core > model 하위 패키지 diary와 user로 2차 분류
