+++ core/ui/src/main/java/com/hilingual/core/ui/model/diary/RecommendExpression.kt
@@ -13,17 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hilingual.core.ui.model
+package com.hilingual.core.ui.model.diary
 
 import androidx.compose.runtime.Immutable
 import kotlinx.collections.immutable.ImmutableList
-import kotlinx.collections.immutable.persistentListOf
 
 @Immutable
-data class DiaryContent(
-    val originalText: String = "",
-    val aiText: String = "",
-    val diffRanges: ImmutableList<Pair<Int, Int>> = persistentListOf(),
-    val imageUrl: String? = null,
-    val isPublished: Boolean = false,
+data class RecommendExpression(
+    val phraseId: Long,
+    val phraseType: ImmutableList<String>,
+    val phrase: String,
+    val explanation: String,
+    val reason: String,
+    val isMarked: Boolean = false,
 )
