+++ core/ui/src/main/java/com/hilingual/core/ui/model/user/NicknameLocalValidationReason.kt
@@ -1,4 +1,4 @@
-package com.hilingual.core.ui.model
+package com.hilingual.core.ui.model.user
 
 enum class NicknameLocalValidationReason {
     TOO_SHORT,
