+++ presentation/mypage/src/main/java/com/hilingual/presentation/mypage/MyPageViewModel.kt
@@ -21,9 +21,9 @@ import androidx.lifecycle.viewModelScope
 import com.hilingual.core.common.app.DeviceInfoProvider
 import com.hilingual.core.common.extension.onLogFailure
 import com.hilingual.core.common.util.UiState
-import com.hilingual.core.ui.model.NicknameValidationStatus
-import com.hilingual.core.ui.model.NicknameLocalValidation
-import com.hilingual.core.ui.model.NicknameLocalValidationReason
+import com.hilingual.core.ui.model.user.NicknameLocalValidation
+import com.hilingual.core.ui.model.user.NicknameLocalValidationReason
+import com.hilingual.core.ui.model.user.NicknameValidationStatus
 import com.hilingual.core.ui.util.NicknameValidator
 import com.hilingual.data.auth.repository.AuthRepository
 import com.hilingual.data.user.repository.UserRepository
