+++ presentation/signup/src/main/java/com/hilingual/presentation/signup/SignUpViewModel.kt
@@ -19,9 +19,9 @@ import android.net.Uri
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.hilingual.core.common.extension.onLogFailure
-import com.hilingual.core.ui.model.NicknameValidationStatus
-import com.hilingual.core.ui.model.NicknameLocalValidation
-import com.hilingual.core.ui.model.NicknameLocalValidationReason
+import com.hilingual.core.ui.model.user.NicknameLocalValidation
+import com.hilingual.core.ui.model.user.NicknameLocalValidationReason
+import com.hilingual.core.ui.model.user.NicknameValidationStatus
 import com.hilingual.core.ui.util.NicknameValidator
 import com.hilingual.data.onboarding.repository.OnboardingRepository
 import com.hilingual.data.user.model.user.UserProfileModel
